chore(test): all sources moved from src/specs to src
resolves #499
