+++ ee_tests/src/page_objects/app/spaceHeader.ts
