+++ ee_tests/src/page_objects/base.page.ts
