+++ ee_tests/src/page_objects/main_dashboard.page.ts
