+++ ee_tests/src/page_objects/space_dashboard.page.ts
