+++ ee_tests/src/page_objects/user_profile.page.ts
