+++ ee_tests/src/quickstart_analytic.spec.ts
