+++ ee_tests/src/quickstart_che.spec.ts
