+++ ee_tests/src/quickstart_cheterminal.spec.ts
