+++ ee_tests/src/quickstart_pipeline.spec.ts
