+++ ee_tests/src/support/interactions.ts
