+++ ee_tests/src/ui/checkbox.ts
