+++ ee_tests/src/ui/modal_dialog.ts
