+++ ee_tests/src/workshop-setup.spec.ts
